Moved ranking api to a separate module
